Remove `out` modifier from generic parameters of `MessageDef`.
