+++ codegen/plugins/codegen-plugins/src/main/kotlin/io/spine/chords/codegen/plugins/MessageOneofObjectGenerator.kt
@@ -26,7 +26,6 @@
 
 package io.spine.chords.codegen.plugins
 
-import com.squareup.kotlinpoet.AnnotationSpec
 import com.squareup.kotlinpoet.ClassName
 import com.squareup.kotlinpoet.FileSpec
 import com.squareup.kotlinpoet.FunSpec
@@ -131,10 +130,7 @@ internal class MessageOneofObjectGenerator(
                 PropertySpec
                     .builder("fieldMap", fieldMapType, PRIVATE)
                     .addAnnotation(
-                        buildSuppressAnnotation(
-                            "UNCHECKED_CAST",
-                            "RemoveRedundantQualifierName"
-                        )
+                        suppressUncheckedCastAndRedundantQualifier()
                     )
                     .initializer(
                         fieldMapInitializer(
@@ -168,19 +164,6 @@ internal class MessageOneofObjectGenerator(
             .build()
     }
 }
-
-/**
- * Builds `@Suppress` annotation with given [warnings].
- */
-@Suppress("SameParameterValue")
-private fun buildSuppressAnnotation(vararg warnings: String) =
-    AnnotationSpec.builder(Suppress::class.asClassName())
-        .also { builder ->
-            warnings.forEach { warning ->
-                builder.addMember("%S", warning)
-            }
-        }
-        .build()
 
 /**
  * Generates initialization code for the `fieldMap` property
