+++ codegen/runtime/src/main/kotlin/io/spine/chords/runtime/MessageDef.kt
@@ -51,7 +51,7 @@ public interface MessageDef<T : Message> {
     /**
      * Returns collection of [MessageField]s generated for the fields of [T] Proto message.
      */
-    public val fields: Collection<MessageField<T, out MessageFieldValue>>
+    public val fields: Collection<MessageField<T, MessageFieldValue>>
 
     /**
      * Returns collection of [MessageOneof]s generated for the oneofs of [T] Proto message.
